+++ frontend/src/main.tsx
@@ -2,11 +2,11 @@ import "./libs/sentry/initializeSentry";
 import { Global } from "@emotion/react";
 import { ErrorBoundary as SentryErrorBoundary } from "@sentry/react";
 import { FallbackErrorUi } from "@shared/components/FallbackErrorUi/FallbackErrorUi";
-import { ToastContainer } from "@shared/components/Toast/components/ToastContainer/ToastContainer";
 import {
   TOAST_DEFAULT_POSITION,
   TOAST_LIMIT,
 } from "@shared/components/Toast/constants/toast.constants";
+import { ToastContainer } from "@shared/components/Toast/ToastContainer/ToastContainer";
 import {
   MutationCache,
   QueryCache,
